+++ engine/src/player/player.cpp
@@ -8,9 +8,9 @@
 /*          Player          */
 /*--------------------------*/
 
-    void Player::update(int deltatime, GCMS& gcm, int dx, int dy, bool slow, bool shot) {
+    void Player::update(float deltatime, GCMS& gcm, int dx, int dy, bool slow, bool shot) {
         auto& v = moveTable[dy + 1][dx + 1];
-        float dtSec = static_cast<float>(deltatime) / 1000.0f;
+        float dtSec = deltatime / 1000.0f;
         float cx = v[0] * speed * (slow ? 0.5f : 1.0f) * dtSec;
         float cy = v[1] * speed * (slow ? 0.5f : 1.0f) * dtSec;
         pos.x += cx;
